WIP: Enable bulk remove/replace blank topic subdivisions
